tools/ceph-dencoder: extract denc_plugin.h out
to prepare for the upcoming change to hold the encoders in DencoderPlugin
